LifeLog/requirements.txt
asgiref==3.8.1
Django==5.0.6
pip==23.0.1
setuptools==65.5.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
django_tables2
django_filter